# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Google Meet BIM Generator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface GenerateBimRequest {
  /** Title or name of the meeting */
  meetingTitle?: string;
  /** Full meeting notes text to analyze */
  meetingNotes: string;
  /** Date of the meeting (ISO format) */
  meetingDate?: string;
}

export type GenerateBimResponseStatus =
  (typeof GenerateBimResponseStatus)[keyof typeof GenerateBimResponseStatus];

export const GenerateBimResponseStatus = {
  pending: "pending",
  processing: "processing",
  completed: "completed",
  failed: "failed",
} as const;

export interface GenerateBimResponse {
  jobId: string;
  status: GenerateBimResponseStatus;
  message?: string;
}

export type BimJobStatus = (typeof BimJobStatus)[keyof typeof BimJobStatus];

export const BimJobStatus = {
  pending: "pending",
  processing: "processing",
  completed: "completed",
  failed: "failed",
} as const;

/**
 * Summary of extracted BIM elements
 */
export type BimJobExtractedElements = {
  spaces?: number;
  walls?: number;
  doors?: number;
  windows?: number;
  columns?: number;
  beams?: number;
  slabs?: number;
};

export type ProcessingStepStatus =
  (typeof ProcessingStepStatus)[keyof typeof ProcessingStepStatus];

export const ProcessingStepStatus = {
  pending: "pending",
  active: "active",
  completed: "completed",
  failed: "failed",
} as const;

export interface ProcessingStep {
  name: string;
  status: ProcessingStepStatus;
  description?: string;
}

export interface BimJob {
  id: string;
  meetingTitle?: string;
  meetingDate?: string;
  status: BimJobStatus;
  /** Progress percentage 0-100 */
  progress?: number;
  /** Current processing step description */
  currentStep?: string;
  steps?: ProcessingStep[];
  errorMessage?: string;
  createdAt: string;
  completedAt?: string;
  /** Size of generated IFC file in bytes */
  fileSize?: number;
  /** Summary of extracted BIM elements */
  extractedElements?: BimJobExtractedElements;
}

export interface BimJobList {
  jobs: BimJob[];
}

export interface ErrorResponse {
  error: string;
  message?: string;
}