# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Google Meet BIM Generator API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Analyzes meeting notes using AI and generates an IFC file
 * @summary Generate IFC BIM file from meeting notes
 */
export const GenerateBimBody = zod.object({
  meetingTitle: zod
    .string()
    .optional()
    .describe("Title or name of the meeting"),
  meetingNotes: zod.string().describe("Full meeting notes text to analyze"),
  meetingDate: zod
    .string()
    .optional()
    .describe("Date of the meeting (ISO format)"),
});

export const GenerateBimResponse = zod.object({
  jobId: zod.string(),
  status: zod.enum(["pending", "processing", "completed", "failed"]),
  message: zod.string().optional(),
});

/**
 * @summary Get status of a BIM generation job
 */
export const GetBimJobParams = zod.object({
  jobId: zod.coerce.string(),
});

export const GetBimJobResponse = zod.object({
  id: zod.string(),
  meetingTitle: zod.string().optional(),
  meetingDate: zod.string().optional(),
  status: zod.enum(["pending", "processing", "completed", "failed"]),
  progress: zod.number().optional().describe("Progress percentage 0-100"),
  currentStep: zod
    .string()
    .optional()
    .describe("Current processing step description"),
  steps: zod
    .array(
      zod.object({
        name: zod.string(),
        status: zod.enum(["pending", "active", "completed", "failed"]),
        description: zod.string().optional(),
      }),
    )
    .optional(),
  errorMessage: zod.string().optional(),
  createdAt: zod.string(),
  completedAt: zod.string().optional(),
  fileSize: zod
    .number()
    .optional()
    .describe("Size of generated IFC file in bytes"),
  extractedElements: zod
    .object({
      spaces: zod.number().optional(),
      walls: zod.number().optional(),
      doors: zod.number().optional(),
      windows: zod.number().optional(),
      columns: zod.number().optional(),
      beams: zod.number().optional(),
      slabs: zod.number().optional(),
    })
    .optional()
    .describe("Summary of extracted BIM elements"),
});

/**
 * @summary Download the generated IFC file
 */
export const DownloadBimFileParams = zod.object({
  jobId: zod.coerce.string(),
});

/**
 * @summary List all BIM generation jobs
 */
export const ListBimJobsResponse = zod.object({
  jobs: zod.array(
    zod.object({
      id: zod.string(),
      meetingTitle: zod.string().optional(),
      meetingDate: zod.string().optional(),
      status: zod.enum(["pending", "processing", "completed", "failed"]),
      progress: zod.number().optional().describe("Progress percentage 0-100"),
      currentStep: zod
        .string()
        .optional()
        .describe("Current processing step description"),
      steps: zod
        .array(
          zod.object({
            name: zod.string(),
            status: zod.enum(["pending", "active", "completed", "failed"]),
            description: zod.string().optional(),
          }),
        )
        .optional(),
      errorMessage: zod.string().optional(),
      createdAt: zod.string(),
      completedAt: zod.string().optional(),
      fileSize: zod
        .number()
        .optional()
        .describe("Size of generated IFC file in bytes"),
      extractedElements: zod
        .object({
          spaces: zod.number().optional(),
          walls: zod.number().optional(),
          doors: zod.number().optional(),
          windows: zod.number().optional(),
          columns: zod.number().optional(),
          beams: zod.number().optional(),
          slabs: zod.number().optional(),
        })
        .optional()
        .describe("Summary of extracted BIM elements"),
    }),
  ),
});