# lib/api-zod/src/generated/types/bimJob.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Google Meet BIM Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { BimJobExtractedElements } from "./bimJobExtractedElements";
import type { BimJobStatus } from "./bimJobStatus";
import type { ProcessingStep } from "./processingStep";

export interface BimJob {
  id: string;
  meetingTitle?: string;
  meetingDate?: string;
  status: BimJobStatus;
  /** Progress percentage 0-100 */
  progress?: number;
  /** Current processing step description */
  currentStep?: string;
  steps?: ProcessingStep[];
  errorMessage?: string;
  createdAt: string;
  completedAt?: string;
  /** Size of generated IFC file in bytes */
  fileSize?: number;
  /** Summary of extracted BIM elements */
  extractedElements?: BimJobExtractedElements;
}