# lib/api-zod/src/generated/types/bimJobList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Google Meet BIM Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { BimJob } from "./bimJob";

export interface BimJobList {
  jobs: BimJob[];
}