# lib/api-zod/src/generated/types/generateBimRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Google Meet BIM Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateBimRequest {
  /** Title or name of the meeting */
  meetingTitle?: string;
  /** Full meeting notes text to analyze */
  meetingNotes: string;
  /** Date of the meeting (ISO format) */
  meetingDate?: string;
}