# lib/api-zod/src/generated/types/generateBimResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Google Meet BIM Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { GenerateBimResponseStatus } from "./generateBimResponseStatus";

export interface GenerateBimResponse {
  jobId: string;
  status: GenerateBimResponseStatus;
  message?: string;
}