# lib/api-zod/src/generated/types/processingStep.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Google Meet BIM Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { ProcessingStepStatus } from "./processingStepStatus";

export interface ProcessingStep {
  name: string;
  status: ProcessingStepStatus;
  description?: string;
}